+++ sparkle/CLI/run_ablation.py
@@ -12,7 +12,7 @@ from sparkle.CLI.help import global_variables as gv
 from sparkle.CLI.help import logging as sl
 from sparkle.platform.settings_objects import Settings, SettingState
 from sparkle.solver import Solver
-from sparkle.instance import instance_set
+from sparkle.instance import Instance_Set
 from sparkle.CLI.help import argparse_custom as ac
 from sparkle.platform import CommandName, COMMAND_DEPENDENCIES
 from sparkle.CLI.initialise import check_for_initialise
@@ -76,11 +76,11 @@ def main(argv: list[str]) -> None:
             args.target_cutoff_time, SettingState.CMD_LINE
         )
     if ac.set_by_user(args, "wallclock_time"):
-        gv.settings().set_config_wallclock_time(
+        gv.settings().set_smac2_wallclock_time(
             args.wallclock_time, SettingState.CMD_LINE
         )
     if ac.set_by_user(args, "number_of_runs"):
-        gv.settings().set_config_number_of_runs(
+        gv.settings().set_configurator_number_of_runs(
             args.number_of_runs, SettingState.CMD_LINE
         )
     if ac.set_by_user(args, "racing"):
@@ -103,21 +103,26 @@ def main(argv: list[str]) -> None:
     instance_set_train = resolve_object_name(
         args.instance_set_train,
         gv.file_storage_data_mapping[gv.instances_nickname_path],
-        gv.settings().DEFAULT_instance_dir, instance_set)
+        gv.settings().DEFAULT_instance_dir, Instance_Set)
     instance_set_test = resolve_object_name(
         args.instance_set_test,
         gv.file_storage_data_mapping[gv.instances_nickname_path],
-        gv.settings().DEFAULT_instance_dir, instance_set)
+        gv.settings().DEFAULT_instance_dir, Instance_Set)
 
     configurator = gv.settings().get_general_sparkle_configurator()
-    configurator.set_scenario_dirs(solver, instance_set_train)
+    config_scenario = configurator.scenario_class.find_scenario(
+        configurator.output_path, solver, instance_set_train)
+    if config_scenario is None:
+        print("No configuration scenario found for combination:\n"
+              f"{configurator.name} {solver.name} {instance_set_train.name}")
+        sys.exit(-1)
     if instance_set_test is None:
         instance_set_test = instance_set_train
 
-    if not configurator.scenario.result_directory.is_dir():
+    if not config_scenario.result_directory.is_dir():
         print("Error: No configuration results found for the given solver and training"
-              " instance set. Ablation needs to have a target configuration.")
-        print("Please run configuration first")
+              " instance set. Ablation needs to have a target configuration. "
+              "Please finish configuration first.")
         sys.exit(-1)
     else:
         print("Configuration exists!")