+++ sparkle/CLI/validate_configured_vs_default.py
@@ -15,7 +15,7 @@ from sparkle.CLI.help.reporting_scenario import Scenario
 from sparkle.configurator.configurator import Configurator
 from sparkle.solver.validator import Validator
 from sparkle.solver import Solver
-from sparkle.instance import instance_set
+from sparkle.instance import Instance_Set
 from sparkle.platform import CommandName, COMMAND_DEPENDENCIES
 from sparkle.CLI.initialise import check_for_initialise
 from sparkle.CLI.help.nicknames import resolve_object_name
@@ -93,19 +93,24 @@ def main(argv: list[str]) -> None:
     instance_set_train = resolve_object_name(
         args.instance_set_train,
         gv.file_storage_data_mapping[gv.instances_nickname_path],
-        gv.settings().DEFAULT_instance_dir, instance_set)
+        gv.settings().DEFAULT_instance_dir, Instance_Set)
     instance_set_test = resolve_object_name(
         args.instance_set_test,
         gv.file_storage_data_mapping[gv.instances_nickname_path],
-        gv.settings().DEFAULT_instance_dir, instance_set)
+        gv.settings().DEFAULT_instance_dir, Instance_Set)
 
     # Make sure configuration results exist before trying to work with them
     configurator = gv.settings().get_general_sparkle_configurator()
-    configurator.set_scenario_dirs(solver, instance_set_train)
-    objectives = gv.settings().get_general_sparkle_objectives()
+    config_scenario = configurator.scenario_class.find_scenario(
+        configurator.output_path, solver, instance_set_train)
+
+    if config_scenario is None:
+        print("No configuration scenario found for combination:\n"
+              f"{configurator.name} {solver.name} {instance_set_train.name}")
+        sys.exit(-1)
+
     # Record optimised configuration
-    _, opt_config_str = configurator.get_optimal_configuration(
-        solver, instance_set_train, objectives[0])
+    _, opt_config_str = configurator.get_optimal_configuration(config_scenario)
     opt_config = Solver.config_str_to_dict(opt_config_str)
 
     validator = Validator(gv.settings().DEFAULT_validation_output, sl.caller_log_dir)
@@ -116,7 +121,7 @@ def main(argv: list[str]) -> None:
         solvers=[solver] * 2,
         configurations=[None, opt_config],
         instance_sets=all_validation_instances,
-        objectives=objectives,
+        objectives=config_scenario.sparkle_objectives,
         cut_off=gv.settings().get_general_target_cutoff_time(),
         sbatch_options=gv.settings().get_slurm_extra_options(as_args=True),
         run_on=run_on)